requirements.txt
python-dotenv
opencv-python
mediapipe
requests
